--- src/main/java/com/rocket/rocket/domain/ClassDto.java
@@ -1,31 +1,0 @@
-package com.rocket.rocket.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ClassDto {
-
-	private String class_num;
-	
-	private String title;
-	
-	private String intro;
-	
-	private String regDate;
-	
-	private String modiDate;
-	
-	private String fileName;
-	
-	private byte fileType;
-	
-	private String filePath;
-	
-	private String thumName;
-	
-	private byte thumType;
-	
-	private String thumPath;
-}
