--- src/main/java/com/rocket/rocket/domain/DoExDto.java
@@ -1,19 +1,0 @@
-package com.rocket.rocket.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class DoExDto {
-
-	private String doex_num;
-	
-	private byte content;
-	
-	private int score;
-	
-	private String ex_num;
-	
-	private String user_num;
-}
