--- src/main/java/com/rocket/rocket/domain/DoHwDto.java
@@ -1,25 +1,0 @@
-package com.rocket.rocket.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class DoHwDto {
-
-	private String dohw_num;
-	
-	private String regDate;
-	
-	private String modiDate;
-	
-	private String fileName;
-	
-	private byte fileType;
-	
-	private String filePath;
-	
-	private String user_num;
-	
-	private String hw_num;
-}
