+++ src/main/java/com/rocket/rocket/domain/DoHwVO.java
@@ -1,4 +1,4 @@
-package com.rocket.rocket.dto;
+package com.rocket.rocket.domain;
 
 import java.util.Date;
