--- src/main/java/com/rocket/rocket/domain/HwDto.java
@@ -1,23 +1,0 @@
-package com.rocket.rocket.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class HwDto {
-
-	private String hw_num;
-	
-	private String title;
-	
-	private String content;
-	
-	private String fileName;
-	
-	private byte fileType;
-	
-	private String filePath;
-	
-	private String lec_num;
-}
