+++ src/main/java/com/rocket/rocket/domain/MakeVO.java
@@ -1,11 +1,9 @@
 package com.rocket.rocket.domain;
 
-import lombok.Getter;
-import lombok.Setter;
+import lombok.Data;
 
-@Getter
-@Setter
-public class MakeDto {
+@Data
+public class MakeVO {
 
 	private String make_num;
