--- src/main/java/com/rocket/rocket/dto/ClassCtgyVO.java
@@ -1,13 +1,0 @@
-package com.rocket.rocket.dto;
-
-import lombok.Data;
-
-@Data
-public class ClassCtgyVO {
-
-	private String class_ctgy_num;
-	
-	private String class_num;
-	
-	private String ctgy_title;
-}
