--- src/main/java/com/rocket/rocket/dto/CtgyVO.java
@@ -1,11 +1,0 @@
-package com.rocket.rocket.dto;
-
-import lombok.Data;
-
-@Data
-public class CtgyVO {
-
-	private String ctgy_title;
-	
-	private String description;
-}
