--- src/main/java/com/rocket/rocket/dto/ExVO.java
@@ -1,23 +1,0 @@
-package com.rocket.rocket.dto;
-
-import java.util.Date;
-
-import lombok.Data;
-
-@Data
-public class ExVO {
-
-	private String ex_num;
-	
-	private String title;
-	
-	private String content;
-	
-	private byte answer;
-	
-	private Date modiDate;
-	
-	private Date regDate;
-	
-	private String lec_num;
-}
