--- src/main/java/com/rocket/rocket/dto/MakeVO.java
@@ -1,13 +1,0 @@
-package com.rocket.rocket.dto;
-
-import lombok.Data;
-
-@Data
-public class MakeVO {
-
-	private String make_num;
-	
-	private String user_num;
-	
-	private String class_num;
-}
